Move list-objects to list-unpacked, now just walks the .git/objects/??/ directories looking for objects and filtering by type.  Use this to do archiving of objects and repacking.
Not sure that walking the .git/objects/??/ directories is the best way to
do this, but it might be.  It is, afterall, exactly what I want to do,
move the unpacked blobs and prune, and pack the unpacked other objects,
and I want to do both of those regardless of the reference graph.